backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
google-genai
python-multipart
aiosmtplib
imap-tools
python-dotenv
